Implementação piloto da estrutura com partial classes. Aplicação de novas interfaces de controle. Implementação para Bradesco e Banco do Brasil (inicial)
