+++ BoletoNetCore/Banco/Bradesco/BancoBradesco.cs
@@ -1,0 +1,41 @@
+using System;
+using System.Collections.Generic;
+using BoletoNetCore.Exceptions;
+using BoletoNetCore.Extensions;
+using static System.String;
+
+namespace BoletoNetCore
+{
+    internal sealed partial class BancoBradesco : BancoFebraban<BancoBradesco> , IBanco
+    {
+        public BancoBradesco()
+        {
+            Codigo = 237;
+            Nome = "Bradesco";
+            Digito = "2";
+            IdsRetornoCnab400RegistroDetalhe = new List<string> { "1" };
+            RemoveAcentosArquivoRemessa = true;
+        }
+
+        public override string FormatarNomeArquivoRemessa(int numeroSequencial)
+        {
+            return $"CB{DateTime.Now.Date.Day:00}{DateTime.Now.Date.Month:00}{numeroSequencial.ToString().PadLeft(9, '0').Right(2)}.rem";
+        }
+
+        public void FormataCedente()
+        {
+            var contaBancaria = Cedente.ContaBancaria;
+
+            if (!CarteiraFactory<BancoBradesco>.CarteiraEstaImplementada(contaBancaria.CarteiraComVariacaoPadrao))
+                throw BoletoNetCoreException.CarteiraNaoImplementada(contaBancaria.CarteiraComVariacaoPadrao);
+
+            contaBancaria.FormatarDados("PAGÁVEL PREFERENCIALMENTE NA REDE BRADESCO OU BRADESCO EXPRESSO.", "", "", 7);
+
+            var codigoCedente = Cedente.Codigo;
+            Cedente.Codigo = codigoCedente.Length <= 20 ? codigoCedente.PadLeft(20, '0') : throw BoletoNetCoreException.CodigoCedenteInvalido(codigoCedente, 20);
+
+            Cedente.CodigoFormatado = $"{contaBancaria.Agencia}-{contaBancaria.DigitoAgencia} / {contaBancaria.Conta}-{contaBancaria.DigitoConta}";
+        }
+
+    }
+}
