+++ BoletoNetCore/Banco/IBanco.cs
@@ -4,6 +4,9 @@ namespace BoletoNetCore
 {
     public interface IBanco
     {
+        /// <summary>
+        /// Beneficiário de uma Cobrança
+        /// </summary>
         Cedente Cedente { get; set; }
         int Codigo { get; }
         string Nome { get; }
@@ -37,6 +40,7 @@ namespace BoletoNetCore
         /// Gera o Trailer do arquivo de remessa
         /// </summary>
         string GerarDetalheRemessa(TipoArquivo tipoArquivo, Boleto boleto, ref int numeroRegistro);
+
         /// <summary>
         /// Gera o Trailer do arquivo de remessa
         /// </summary>
@@ -47,14 +51,64 @@ namespace BoletoNetCore
                                             int numeroRegistroCobrancaCaucionada, decimal valorCobrancaCaucionada,
                                             int numeroRegistroCobrancaDescontada, decimal valorCobrancaDescontada);
 
-        void LerHeaderRetornoCNAB240(ArquivoRetorno arquivoRetorno, string registro);
-        void LerDetalheRetornoCNAB240SegmentoT(ref Boleto boleto, string registro);
-        void LerDetalheRetornoCNAB240SegmentoU(ref Boleto boleto, string registro);
+        string FormatarNomeArquivoRemessa(int numeroSequencial);
+    }
+
+    /// <summary>
+    /// Implementa Remessa e Tetorno de Cobrança no Formato CNAB400
+    /// </summary>
+    public interface IBancoCNAB400 : IBanco
+    {
+        //remessa
+        string GerarHeaderRemessaCNAB400(ref int numeroArquivoRemessa, ref int numeroRegistroGeral);
+        string GerarDetalheRemessaCNAB400(Boleto boleto, ref int registro);
+        string GerarTrailerRemessaCNAB400(ref int numeroArquivoRemessa);
+
+        //retorno
         void LerHeaderRetornoCNAB400(string registro);
         void LerDetalheRetornoCNAB400Segmento1(ref Boleto boleto, string registro);
         void LerDetalheRetornoCNAB400Segmento7(ref Boleto boleto, string registro);
         void LerTrailerRetornoCNAB400(string registro);
+    }
 
-        string FormatarNomeArquivoRemessa(int numeroSequencial);
+    /// <summary>
+    /// Implementa Remessa e Retorno de Cobrança no formato CNAB240 em uma Intituição Financeira
+    /// </summary>
+    public interface IBancoCNAB240 : IBanco
+    {
+        /// <summary>
+        /// 1 - Header de Remessa e Lote do Arquivo de Remessa
+        /// </summary>
+        /// <param name="numeroArquivoRemessa"></param>
+        /// <param name="numeroRegistro"></param>
+        /// <returns></returns>
+        string GerarHeaderRemessaCNAB240(ref int numeroArquivoRemessa, ref int numeroRegistro);
+        string GerarHeaderLoteRemessaCNAB240(ref int numeroArquivoRemessa, ref int numeroRegistro);
+        string GerarDetalheRemessaCNAB240(Boleto boleto, ref int registro);
+        string GerarTrailerLoteRemessaCNAB240(ref int numeroArquivoRemessa, int numeroRegistroGeral,
+            int numeroRegistroCobrancaSimples, decimal valorCobrancaSimples,
+            int numeroRegistroCobrancaVinculada, decimal valorCobrancaVinculada,
+            int numeroRegistroCobrancaCaucionada, decimal valorCobrancaCaucionada,
+            int numeroRegistroCobrancaDescontada, decimal valorCobrancaDescontada);
+
+        string GerarTrailerRemessaCNAB240(int numeroRegistroGeral,
+            int numeroRegistroCobrancaSimples, decimal valorCobrancaSimples,
+            int numeroRegistroCobrancaVinculada, decimal valorCobrancaVinculada,
+            int numeroRegistroCobrancaCaucionada, decimal valorCobrancaCaucionada,
+            int numeroRegistroCobrancaDescontada, decimal valorCobrancaDescontada);
+
+        void LerHeaderRetornoCNAB240(ArquivoRetorno arquivoRetorno, string registro);
+        void LerDetalheRetornoCNAB240SegmentoT(ref Boleto boleto, string registro);
+        void LerDetalheRetornoCNAB240SegmentoU(ref Boleto boleto, string registro);
+    }
+
+    /// <summary>
+    /// Implementa Registro Online de Boleto
+    /// </summary>
+    public interface IBancoOnlineRest : IBanco
+    {
+        string GerarToken();
+        void RegistrarBoleto(ref Boleto boleto, string registro);
+        //StatusBoletoOnline ConsultarStatus(ref Boleto boleto, string registro);
     }
 }
